Hisilicon/D05: add cpld to get sfp status
This patch is for https://bugs.linaro.org/show_bug.cgi?id=2711.

When there is no optical module plugged in XGE port, the interfaces (ethX)
will show up/down information on console, such as below:

[root@localhost ~]# [  676.034368] hns-nic HISI00C2:03 eth3: link up
[  676.038840] IPv6: ADDRCONF(NETDEV_CHANGE): eth3: link becomes ready
[  677.058362] hns-nic HISI00C2:03 eth3: link down
[  692.418366] hns-nic HISI00C2:03 eth3: link up
[  693.442359] hns-nic HISI00C2:03 eth3: link down
[  748.738365] hns-nic HISI00C2:03 eth3: link up
[  749.762363] hns-nic HISI00C2:03 eth3: link down

This patch adds a logic to check whether the optical module is absent or not,
if absent, means the interface is down, otherwise check the mac link status.

Signed-off-by: Daode Huang <huangdaode@hisilicon.com>